backend/requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
djangorestframework>=3.15
django-cors-headers>=4.3
gunicorn>=21.2
whitenoise>=6.6
python-dotenv>=1.0
django-jazzmin>=3.0
